(#66) Extract metricbeat test logic out of the test runner
That will allow adding new future implementations, such as filebeat
